% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{alphaPrior<-,TopicExperiment-method}
\alias{alphaPrior<-,TopicExperiment-method}
\title{Set Alpha Prior Matrix}
\usage{
\S4method{alphaPrior}{TopicExperiment}(te) <- value
}
\arguments{
\item{te}{A TopicExperiment object.}

\item{value}{The new alpha prior matrix.}
}
\value{
The updated TopicExperiment object.
}
\description{
Sets the alpha prior matrix (cell-by-topic) for a TopicExperiment object.
}
