% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{alphaPrior,TopicExperiment-method}
\alias{alphaPrior,TopicExperiment-method}
\title{Get Alpha Prior Matrix}
\usage{
\S4method{alphaPrior}{TopicExperiment}(te)
}
\arguments{
\item{te}{A TopicExperiment object.}
}
\value{
The alpha prior matrix.
}
\description{
Retrieves the alpha prior matrix (cell-by-topic) from a TopicExperiment object.
}
