% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{betaPrior<-,TopicExperiment-method}
\alias{betaPrior<-,TopicExperiment-method}
\title{Set Beta Prior Matrix}
\usage{
\S4method{betaPrior}{TopicExperiment}(te) <- value
}
\arguments{
\item{te}{A TopicExperiment object.}

\item{value}{The new beta prior matrix.}
}
\value{
The updated TopicExperiment object.
}
\description{
Sets the beta prior matrix (gene-by-topic) for a TopicExperiment object.
}
