% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{betaPrior,TopicExperiment-method}
\alias{betaPrior,TopicExperiment-method}
\title{Get Beta Prior Matrix}
\usage{
\S4method{betaPrior}{TopicExperiment}(te)
}
\arguments{
\item{te}{A TopicExperiment object.}
}
\value{
The beta prior matrix.
}
\description{
Retrieves the beta prior matrix (gene-by-topic) from a TopicExperiment object.
}
