% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{build_alpha}
\alias{build_alpha}
\title{Build Alpha Prior Matrix}
\usage{
build_alpha(
  te,
  labels = NULL,
  K = 10,
  balanced = FALSE,
  disease = NULL,
  disease_type = NULL
)
}
\arguments{
\item{te}{A SingleCellTopicExperiment or SpatialTopicExperiment object.}

\item{labels}{Character string specifying the column name in `colData(te)`
that contains labels for guiding topic assignment.}

\item{K}{Integer specifying the number of topics (ignored if labels are provided).}
}
\value{
A matrix representing the alpha prior.
}
\description{
Constructs the alpha prior matrix for topic modeling. If labels are provided,
it creates a sparse matrix based on the provided labels; otherwise, it assigns
a default Dirichlet prior.
}
