% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{build_beta}
\alias{build_beta}
\title{Build Beta Prior Matrix}
\usage{
build_beta(te, K)
}
\arguments{
\item{te}{A SingleCellTopicExperiment or SpatialTopicExperiment object.}

\item{K}{Integer specifying the number of topics.}
}
\value{
A matrix representing the beta prior.
}
\description{
Constructs the beta prior matrix for topic modeling, assigning a default prior.
}
