% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{build_SufStats}
\alias{build_SufStats}
\title{Initialize Sufficient Statistics for Topic Modeling}
\usage{
build_SufStats(te, K)
}
\arguments{
\item{te}{A SingleCellTopicExperiment or SpatialTopicExperiment object.}

\item{K}{Integer specifying the number of topics.}
}
\value{
The input `te` object with initialized sufficient statistics.
}
\description{
Initializes the sufficient statistics matrices (`ndk` and `nwk`) with zeros.
}
