% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BuildDiseaseAlpha}
\alias{BuildDiseaseAlpha}
\title{Build Disease-informed alpha priors}
\usage{
BuildDiseaseAlpha(a, disease)
}
\arguments{
\item{a}{A numeric matrix representing the original alpha priors.}

\item{disease}{An unsigned integer vector indicating disease presence (1) or absence (0) for each row.}
}
\value{
A numeric matrix with adjusted alpha values. Rows are modified based on the corresponding disease indicator:
- If `disease[cur] == 1`, the row is element-wise multiplied by {0.5, 0.5}.
- If `disease[cur] == 0`, the row is element-wise multiplied by {0.9, 0.1}.
}
\description{
updates the initialized label-guided alpha prior matrix to include disease information. We assume that most cells from healthy patients are healthy and that there is an equal chance of a cell from a diseased patient representing a healthy or diseased state.
}
\examples{
\dontrun{
a <- matrix(c(1, 2, 3, 4), nrow = 2, byrow = TRUE)
disease <- c(1, 0)
BuildDiseaseAlpha(a, disease)
}
}
