% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{buildPhi}
\alias{buildPhi}
\title{Compute Phi Matrix (Gene-Topic Distributions)}
\usage{
buildPhi(te)
}
\arguments{
\item{te}{A SingleCellTopicExperiment or SpatialTopicExperiment object.}
}
\value{
The input `te` object with the `phi` matrix computed.
}
\description{
Computes the phi matrix, representing the per-gene topic distributions.
}
