% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{buildPriors}
\alias{buildPriors}
\title{Build Priors for Topic Modeling}
\usage{
buildPriors(te, labels = NULL, guided = FALSE, K = 10)
}
\arguments{
\item{te}{A SingleCellTopicExperiment or SpatialTopicExperiment object.}

\item{labels}{Character string specifying the column name in `colData(te)` for topic guidance.}

\item{guided}{Logical indicating whether topic modeling should be guided by labels.}

\item{K}{Integer specifying the number of topics (ignored if guided is TRUE and labels are provided).}
}
\value{
The input `te` object with alpha and beta prior matrices initialized.
}
\description{
Initializes the alpha and beta prior matrices for a topic modeling experiment.
}
