% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{buildTheta}
\alias{buildTheta}
\title{Compute Theta Matrix (Cell-Topic Distributions)}
\usage{
buildTheta(te)
}
\arguments{
\item{te}{A SingleCellTopicExperiment or SpatialTopicExperiment object.}
}
\value{
The input `te` object with the `theta` matrix computed.
}
\description{
Computes the theta matrix, representing the per-cell topic distributions.
}
