% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTM.R
\name{clust_optim}
\alias{clust_optim}
\title{Optimize clustering resolution to achieve a target number of clusters}
\usage{
clust_optim(
  spe,
  target_clusters,
  max_iterations = 100,
  X = "adj",
  k = 20,
  alg = "louvain"
)
}
\arguments{
\item{spe}{A `SpatialTopicExperiment` object.}

\item{target_clusters}{The desired number of clusters to achieve.}

\item{max_iterations}{The maximum number of iterations to run for adjusting the resolution (default is 100).}

\item{X}{The name of the reduced dimension to use for clustering. Default is `'adj'`, which assumes an adjacency matrix is used for clustering.}

\item{k}{The number of nearest neighbors to consider when constructing the graph (default is 20).}

\item{alg}{The algorithm to use for clustering. The default is `'louvain'`, but other algorithms can be used (e.g., `'walktrap'`).}
}
\value{
A list with two elements:
  \describe{
    \item{clustering}{A vector of cluster assignments for each spot.}
    \item{resolution}{The resolution parameter used to achieve the target number of clusters.}
  }
}
\description{
This function performs clustering optimization by adjusting the resolution parameter in the Louvain algorithm until the target number of clusters is reached. It iteratively adjusts the resolution parameter, running clustering with each iteration, until the desired number of clusters is obtained or the maximum number of iterations is reached.
}
\details{
The function uses an adjacency matrix to construct a graph and performs clustering on the reduced-dimensional representation of the spatial experiment object. The algorithm used for clustering is Louvain by default, but other algorithms can be specified.
}
