% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STM.R
\name{fit_mlp}
\alias{fit_mlp}
\title{Fit MLP Model}
\usage{
fit_mlp(mlp, train_dl, lr, max_epoch, class_prop)
}
\arguments{
\item{mlp}{An MLP model (of class `nn_module`).}

\item{train_dl}{A data loader for the training data.}

\item{lr}{Learning rate for the optimizer.}

\item{max_epoch}{Maximum number of epochs for training.}

\item{class_prop}{Class proportions for weighting the loss function.}
}
\value{
The trained MLP model.
}
\description{
This function trains a multi-layer perceptron (MLP) model using the provided training data loader, learning rate, maximum number of epochs, and class proportions.
}
