% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTM.R
\name{get_all_pred}
\alias{get_all_pred}
\title{Get All Predictions from Relational Topic Model (RTM)}
\usage{
get_all_pred(spe, loss_fun)
}
\arguments{
\item{spe}{A `SpatialTopicExperiment` object}

\item{loss_fun}{An integer specifying the loss function to be used for prediction.}
}
\value{
A matrix of predicted values based on the RTM model.
}
\description{
This function computes predictions using the trained RTM weights stored in the metadata of a `SpatialTopicExperiment` object.
}
\details{
This function retrieves the trained RTM weights from `metadata(spe)[['RTM_weights']]` and applies them to
the topic distribution (`theta(spe)`) to compute relational topic predictions.
}
\seealso{
[RTM()] for training the Relational Topic Model.
}
