% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Neighbors.R
\name{get_nbrs}
\alias{get_nbrs}
\title{Get neighboring spots based on spatial distance and grouping}
\usage{
get_nbrs(spe, samples, cell_ids, group_by = NULL, dist = 1, loss_fun = 1)
}
\arguments{
\item{spe}{A `SpatialTopicExperiment` object.}

\item{samples}{The column name in `colData(spe)` representing sample identifiers.}

\item{cell_ids}{The column name in `colData(spe)` representing cell identifiers.}

\item{group_by}{The column name in `colData(spe)` to group the spots by. If `NULL`, no grouping is performed, and neighbors are selected based on distance alone.}

\item{dist}{The maximum spatial distance within which to consider spots as neighbors (default is 1).}

\item{loss_fun}{A numeric value to control the use of loss functions:
- 0: No loss function applied.
- 1: Applies a loss function that includes randomly selected negative neighbors based on spatial distance.
- 2: Applies a Euclidean distance-based loss function.}
}
\value{
A list where each element corresponds to a spot, and contains its neighboring spots, represented by a matrix with indices and weights.
  If a loss function is used, additional information will be appended to the matrix.
}
\description{
This function identifies neighboring spots for each spot in a spatial experiment object based on their spatial coordinates. It can consider spots from the same sample and group, and allows for the inclusion of loss functions to modify neighbor selection.
}
\details{
If `group_by` is provided, the function selects neighbors from the same sample and group, and may sample neighbors based on spatial proximity. If no `group_by` is given, neighbors are selected within a given distance threshold (`dist`) from the current spot, and optionally modified by a loss function.
}
\examples{
# Example 1: Getting neighbors for spots grouped by a specific column
nbr_list <- get_nbrs(spe = spe_object, samples = "sample_id", cell_ids = "cell_id", group_by = "group", dist = 1)

# Example 2: Getting neighbors without grouping, applying a distance threshold and no loss function
nbr_list <- get_nbrs(spe = spe_object, samples = "sample_id", cell_ids = "cell_id", dist = 1, loss_fun = 0)

# Example 3: Using a loss function to sample negative neighbors
nbr_list <- get_nbrs(spe = spe_object, samples = "sample_id", cell_ids = "cell_id", loss_fun = 1)

}
