% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTM.R
\name{GTM}
\alias{GTM}
\title{Guided Topic Model (GTM)}
\usage{
GTM(
  scte,
  K,
  D,
  num_threads = 1,
  maxiter = 100,
  verbal = TRUE,
  zero_gamma = FALSE,
  rand_gamma = TRUE,
  thresh = 1e-08,
  burnin = 1
)
}
\arguments{
\item{scte}{A `SingleCellTopicExperiment` or `SpatialTopicExperiment` object containing count data and prior matrices.}

\item{K}{Integer, the number of topics to infer.}

\item{D}{Integer, the number of cells (documents) in the dataset.}

\item{num_threads}{Integer, the number of threads to use for parallel computation. Default is 1.}

\item{maxiter}{Integer, the maximum number of iterations for training. Default is 100.}

\item{verbal}{Logical, whether to print progress messages. Default is TRUE.}

\item{zero_gamma}{Logical, whether to initialize gamma values to zero. Default is FALSE.}

\item{rand_gamma}{Logical, whether to initialize gamma values randomly. Default is TRUE.}

\item{burnin}{Maximum number of iterations to run during the STM LDA step for each sample (default is 0)}
}
\value{
A `SingleCellTopicExperiment` or `SpatialTopicExperiment` object with updated topic distributions.
}
\description{
This function trains a Guided Topic Model (GTM) using the input TopicExperiment (TE) object.
GTM optimizes topic distributions using prior knowledge provided through metadata labels.
}
