% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTM.R
\name{inferGEx}
\alias{inferGEx}
\title{Cell type-specific Gene Expression Deconvolution}
\usage{
inferGEx(scte, num_threads = 1, maxiter = 50, verbal = TRUE, phi)
}
\arguments{
\item{scte}{A `SingleCellTopicExperiment` or `SpatialTopicExperiment` object containing count data.}

\item{num_threads}{Integer, the number of threads to use for parallel computation. Default is 1.}

\item{maxiter}{Integer, the maximum number of iterations for inference. Default is 50.}

\item{verbal}{Logical, whether to print progress messages. Default is TRUE.}

\item{phi}{A matrix representing the topic-gene distribution from a trained model.}
}
\value{
A list containing the topic distributions and gene counts for each observation
}
\description{
This function infers Cell type-specific gene expression profiles from bulk data using a pre-trained topic model
}
