% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTM.R
\name{inferTopics}
\alias{inferTopics}
\title{Topic Inference using Pretrained Model}
\usage{
inferTopics(scte, num_threads = 1, maxiter = 50, verbal = TRUE, phi)
}
\arguments{
\item{scte}{A `SingleCellTopicExperiment` or `SpatialTopicExperiment` object containing count data.}

\item{num_threads}{Integer, the number of threads to use for parallel computation. Default is 1.}

\item{maxiter}{Integer, the maximum number of iterations for inference. Default is 50.}

\item{verbal}{Logical, whether to print progress messages. Default is TRUE.}

\item{phi}{A matrix representing the topic-gene distribution from a trained model.}
}
\value{
A `SingleCellTopicExperiment` or `SpatialTopicExperiment` object with updated inferred topic distributions.
}
\description{
This function infers topic distributions for new data using a pretrained topic model.
}
