% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTM.R
\name{minmax_norm}
\alias{minmax_norm}
\title{Min-Max Normalization}
\usage{
minmax_norm(adj_mat)
}
\arguments{
\item{adj_mat}{A numeric matrix to be normalized.}
}
\value{
A numeric matrix with values scaled between 0 and 1.
}
\description{
This function performs min-max normalization on an input matrix, scaling the values to be between 0 and 1.
}
\examples{
# Example: Normalize a matrix
mat <- matrix(runif(100, 0, 10), nrow = 10, ncol = 10)
norm_mat <- minmax_norm(mat)
print(norm_mat)

}
