% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STM.R
\name{mlp_data}
\alias{mlp_data}
\title{MLP Data Dataset}
\usage{
mlp_data(x, y)
}
\arguments{
\item{x}{A tensor representing the input data.}

\item{y}{A tensor representing the target data.}
}
\value{
A dataset object for the MLP model.
}
\description{
This function defines a dataset for the MLP model, including methods for initialization, getting batches, and getting the length of the dataset.
}
