% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{ndk<-,TopicExperiment-method}
\alias{ndk<-,TopicExperiment-method}
\title{Set Cell-Topic Matrix (ndk)}
\usage{
\S4method{ndk}{TopicExperiment}(te) <- value
}
\arguments{
\item{te}{A TopicExperiment object.}

\item{value}{The new cell-topic matrix (ndk).}
}
\value{
The updated TopicExperiment object.
}
\description{
Sets the cell-topic matrix (ndk) for a TopicExperiment object.
}
