% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{ndk,TopicExperiment-method}
\alias{ndk,TopicExperiment-method}
\title{Get Cell-Topic Matrix (ndk)}
\usage{
\S4method{ndk}{TopicExperiment}(te)
}
\arguments{
\item{te}{A TopicExperiment object.}
}
\value{
The cell-topic matrix (ndk).
}
\description{
Retrieves the cell-topic matrix (ndk) from a TopicExperiment object.
}
