% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{nwk<-,TopicExperiment-method}
\alias{nwk<-,TopicExperiment-method}
\title{Set Gene-Topic Matrix (nwk)}
\usage{
\S4method{nwk}{TopicExperiment}(te) <- value
}
\arguments{
\item{te}{A TopicExperiment object.}

\item{value}{The new gene-topic matrix (nwk).}
}
\value{
The updated TopicExperiment object.
}
\description{
Sets the gene-topic matrix (nwk) for a TopicExperiment object.
}
