% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{nwk,TopicExperiment-method}
\alias{nwk,TopicExperiment-method}
\title{Get Gene-Topic Matrix (nwk)}
\usage{
\S4method{nwk}{TopicExperiment}(te)
}
\arguments{
\item{te}{A TopicExperiment object.}
}
\value{
The gene-topic matrix (nwk).
}
\description{
Retrieves the gene-topic matrix (nwk) from a TopicExperiment object.
}
