% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{phi<-,TopicExperiment-method}
\alias{phi<-,TopicExperiment-method}
\title{Set Phi Matrix (Gene-Topic Distributions)}
\usage{
\S4method{phi}{TopicExperiment}(te) <- value
}
\arguments{
\item{te}{A TopicExperiment object.}

\item{value}{The new phi matrix.}
}
\value{
The updated TopicExperiment object.
}
\description{
Sets the phi matrix (gene-topic distributions) for a TopicExperiment object.
}
