% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{phi,TopicExperiment-method}
\alias{phi,TopicExperiment-method}
\title{Get Phi Matrix (Gene-Topic Distributions)}
\usage{
\S4method{phi}{TopicExperiment}(te)
}
\arguments{
\item{te}{A TopicExperiment object.}
}
\value{
The phi matrix.
}
\description{
Retrieves the phi matrix (gene-topic distributions) from a TopicExperiment object.
}
