% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STM.R
\name{pred_mlp}
\alias{pred_mlp}
\title{Predict with MLP Model}
\usage{
pred_mlp(mlp, x, device = torch_device("cpu"), dummy_topic = FALSE)
}
\arguments{
\item{mlp}{A trained MLP model (of class `nn_module`).}

\item{x}{A matrix representing the input data.}

\item{device}{The device on which to run the model. Defaults to `"cpu"`.}

\item{dummy_topic}{Logical indicating whether to include a dummy topic in the input (default is `FALSE`).}
}
\value{
A vector of predicted class labels.
}
\description{
This function generates predictions using a trained multi-layer perceptron (MLP) model and the provided input data.
}
