% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTM.R
\name{RTM}
\alias{RTM}
\title{Train a Relational Topic Model (RTM)}
\usage{
RTM(
  scte,
  K,
  nbr_list,
  loss_fun = 1,
  num_threads = 1,
  maxiter = 100,
  lr = 1e-05,
  verbal = TRUE,
  zero_gamma = FALSE,
  rand_gamma = TRUE,
  m_update = T
)
}
\arguments{
\item{scte}{A `SpatialTopicExperiment` object containing the single-cell data.}

\item{K}{An integer specifying the number of topics.}

\item{nbr_list}{A list specifying the neighbors of individual cells.}

\item{loss_fun}{An integer specifying the loss function to be used in RTM training (default: 1).}

\item{num_threads}{An integer indicating the number of threads to use for parallel computation (default: 1).}

\item{maxiter}{An integer specifying the maximum number of training iterations (default: 100).}

\item{verbal}{A logical indicating whether to print progress messages (default: TRUE).}

\item{zero_gamma}{A logical specifying whether to initialize gamma values to zero (default: FALSE).}

\item{rand_gamma}{A logical indicating whether to randomly initialize gamma values (default: TRUE).}
}
\value{
A `SpatialTopicExperiment` object with updated RTM weights stored in `metadata(scte)[['RTM_weights']]`.
}
\description{
This function trains a Relational Topic Model (RTM) on a `SpatialTopicExperiment` object.
The RTM incorporates relational data (e.g., cell-cell interactions) into the topic modeling process.
}
\details{
The RTM extends topic modeling by incorporating relational information, such as neighborhood information.
It uses a loss function specified by `loss_fun` to optimize the topic distributions.
}
\seealso{
[GTM()] for the guided topic model.
}
