% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTM.R
\name{rtm_smooth}
\alias{rtm_smooth}
\title{Smooth labels based on spatial proximity}
\usage{
rtm_smooth(spe, labels, nbr_list, k = NULL)
}
\arguments{
\item{spe}{A `SpatialTopicExperiment` object. Note that `spe` must include a `spatialCoords` slot with array coordinates (`array_row`, `array_col`).}

\item{labels}{A vector of current labels corresponding to each spot in `spe`.}

\item{nbr_list}{A list where each element corresponds to a spot and contains indices of its neighbors.}

\item{k}{An optional integer specifying the number of nearest neighbors to consider when smoothing the labels. If `NULL`, `nbr_list` is used instead.}
}
\value{
A character vector containing the new labels for each spot after smoothing.
}
\description{
This function smooths the labels of spots in a spatial experiment object based on their spatial proximity to other spots. It assigns new labels to spots based on the labels of their neighboring spots.
}
\details{
If `k` is provided, the function computes the spatial distance matrix between spots using their array coordinates (`array_row`, `array_col`) and considers the `k` nearest neighbors for smoothing. Otherwise, the function uses a user-provided list of neighbors for each spot.
}
\examples{
# Example 1: Using k-nearest neighbors
new_labels <- rtm_smooth(spe = spe_object, labels = initial_labels, nbr_list = NULL, k = 5)

# Example 2: Using a pre-computed neighbor list
new_labels <- rtm_smooth(spe = spe_object, labels = initial_labels, nbr_list = precomputed_nbr_list, k = NULL)

}
