% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellTopicExperiment.R
\docType{class}
\name{SingleCellTopicExperiment-class}
\alias{SingleCellTopicExperiment-class}
\title{SingleCellTopicExperiment Class}
\description{
A class that extends `SingleCellExperiment` to incorporate topic modeling
components for single-cell RNA-seq analysis.
}
\section{Slots}{

\describe{
\item{\code{theta}}{A matrix representing cell-topic distributions.}

\item{\code{phi}}{A matrix representing gene-topic distributions.}

\item{\code{alphaPrior}}{A matrix representing prior information for cell-topic distributions.}

\item{\code{betaPrior}}{A matrix representing prior information for gene-topic distributions.}

\item{\code{nwk}}{A matrix storing gene-topic counts.}

\item{\code{ndk}}{A matrix storing cell-topic counts.}
}}

