% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialTopicExperiment.R
\docType{class}
\name{SpatialTopicExperiment-class}
\alias{SpatialTopicExperiment-class}
\title{SpatialTopicExperiment Class and Constructor}
\description{
The `SpatialTopicExperiment` class extends `SpatialExperiment`, incorporating topic modeling
for spatial transcriptomics data. It follows the same logic as `SingleCellTopicExperiment` (@seealso [SingleCellTopicExperiment()]),
adapting it for spatial data.
}
\section{Slots}{

\describe{
\item{\code{theta}}{A matrix storing cell-topic distributions.}

\item{\code{phi}}{A matrix storing gene-topic distributions.}

\item{\code{alphaPrior}}{A matrix storing topic priors per cell.}

\item{\code{betaPrior}}{A matrix storing gene priors per topic.}

\item{\code{nwk}}{A matrix storing topic-gene counts.}

\item{\code{ndk}}{A matrix storing topic-cell counts.}
}}

