% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialTopicExperiment.R
\name{SpatialTopicExperiment}
\alias{SpatialTopicExperiment}
\title{Constructor for SpatialTopicExperiment}
\usage{
SpatialTopicExperiment(
  spe,
  guided = FALSE,
  labels = NULL,
  K = NULL,
  hvg = NULL,
  verbal = FALSE
)
}
\arguments{
\item{spe}{A `SpatialExperiment` object.}

\item{guided}{Logical; if `TRUE`, the initialization is guided by cell labels.}

\item{labels}{A character string specifying the column in `colData(spe)` containing cell labels.}

\item{K}{Number of topics. If `NULL` and `guided` is `FALSE`, defaults to `K = 10`.}

\item{hvg}{Integer specifying the number of highly variable genes to retain.}

\item{verbal}{Logical; if `TRUE`, prints progress messages.}
}
\value{
A `SpatialTopicExperiment` object initialized with topic modeling parameters.
}
\description{
Initializes a `SpatialTopicExperiment` object, setting up topic modeling structures
for spatial transcriptomics data. It follows a similar process as `SingleCellTopicExperiment`.
}
