% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STM.R
\name{STM_Torch_burnin}
\alias{STM_Torch_burnin}
\title{Burn-in Phase for SpaTM-S using Torch}
\usage{
STM_Torch_burnin(
  spe,
  labels,
  num_threads = 1L,
  maxiter = 100L,
  verbal = TRUE,
  zero_gamma = TRUE,
  rand_gamma = FALSE,
  thresh = 1e-05,
  lr = 1e-04,
  mlp,
  mlp_layers,
  mlp_epoch = 1000,
  burnin = 1
)
}
\arguments{
\item{spe}{A `SpatialTopicExperiment` object.}

\item{labels}{A vector of labels for the cells.}

\item{num_threads}{Number of threads to use for parallel computation (default is 1).}

\item{maxiter}{Maximum number of iterations for the burn-in phase (default is 100).}

\item{verbal}{Logical indicating whether to display progress messages (default is `TRUE`).}

\item{zero_gamma}{Logical indicating whether to initialize the gamma parameters to zero (default is `TRUE`).}

\item{rand_gamma}{Logical indicating whether to randomize the gamma parameters (default is `FALSE`).}

\item{thresh}{Convergence threshold for stopping the burn-in phase (default is 0.00001).}

\item{lr}{Learning rate for optimization (default is 0.0001).}

\item{mlp}{A model definition for the multi-layer perceptron (MLP).}

\item{mlp_layers}{A vector specifying the number of units in each layer of the MLP.}

\item{mlp_epoch}{The number of epochs for training the MLP during the burn-in phase.}

\item{burnin}{Maximum number of iterations to run during the STM LDA step for each sample (default is 0)}

\item{spe_val}{A validation set.}
}
\description{
This function performs the burn-in phase for training SpaTM-S using a multi-layer perceptron (MLP) model.
}
\seealso{
\link{STM}
}
