% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STM.R
\name{test_mlp}
\alias{test_mlp}
\title{Testthat MLP forward pass helper function}
\usage{
test_mlp(theta_mat, mlp_layers, mlp_parameters, dummy_topic = FALSE)
}
\arguments{
\item{theta_mat}{input feature matrix}

\item{mlp_layers}{number of MLP hidden layers}

\item{mlp_parameters}{list of MLP parameters}

\item{dummy_topic}{boolean for whether to use dummy topic in model or not}
}
\value{
outputs from the Rcpp forward pass implementation
}
\description{
This helper function is only used in the testing suite to test the Rcpp forward pass
}
\details{
The function supports various parameters for training, including regularization, optimization settings, and the option for a validation set to prevent overfitting in the Torch-STM mode.
}
