% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{theta<-,TopicExperiment-method}
\alias{theta<-,TopicExperiment-method}
\title{Set Theta Matrix (Cell-Topic Distributions)}
\usage{
\S4method{theta}{TopicExperiment}(te) <- value
}
\arguments{
\item{te}{A TopicExperiment object.}

\item{value}{The new theta matrix.}
}
\value{
The updated TopicExperiment object.
}
\description{
Sets the theta matrix (cell-topic distributions) for a TopicExperiment object.
}
