% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopicExperiment.utils.R
\name{theta,TopicExperiment-method}
\alias{theta,TopicExperiment-method}
\title{Get Theta Matrix (Cell-Topic Distributions)}
\usage{
\S4method{theta}{TopicExperiment}(te)
}
\arguments{
\item{te}{A TopicExperiment object.}
}
\value{
The theta matrix.
}
\description{
Retrieves the theta matrix (cell-topic distributions)
from a TopicExperiment object.
}
